core/chat/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.2.1
filelock==3.15.4
fireworks-ai==0.14.0
flask==3.0.3
frozenlist==1.4.1
fsspec==2024.6.0
greenlet==3.0.3
groq==0.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.4
idna==3.7
importlib-metadata==7.2.1
itsdangerous==2.2.0
jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.5
langchain-community==0.2.4
langchain-core==0.2.9
langchain-fireworks==0.1.3
langchain-groq==0.1.5
langchain-huggingface==0.0.3
langchain-postgres==0.0.7
langchain-text-splitters==0.2.1
langsmith==0.1.81
MarkupSafe==2.1.5
marshmallow==3.21.3
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
openai==1.35.3
orjson==3.10.5
packaging==24.1
pgvector==0.2.5
pillow==10.3.0
psycopg==3.1.19
psycopg-binary==3.1.19
psycopg-pool==3.2.2
pydantic==2.7.4
pydantic-core==2.18.4
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==3.0.1
sniffio==1.3.1
SQLAlchemy==2.0.31
sympy==1.12.1
tenacity==8.4.2
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.1
tqdm==4.66.4
transformers==4.41.2
triton==2.3.1
typing-extensions==4.12.2
typing-inspect==0.9.0
urllib3==2.2.2
werkzeug==3.0.3
yarl==1.9.4
zipp==3.19.2
gunicorn==22.0.0